[status_manager] Moved update_db.py up from legacy health manager dir
Add note in update_db.py for clarification and as placeholder for
rework of status manager mechanics.
