--- octavia_f5/controller/statusmanager/legacy_healthmanager/health_drivers/update_base.py
@@ -1,27 +1,0 @@
-# Copyright 2018 GoDaddy
-#
-# Licensed under the Apache License, Version 2.0 (the "License"); you may
-# not use this file except in compliance with the License. You may obtain
-# a copy of the License at
-#
-# http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
-# WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
-# License for the specific language governing permissions and limitations
-# under the License.
-
-import abc
-
-
-class HealthUpdateBase(object):
-    @abc.abstractmethod
-    def update_health(self, health, srcaddr):
-        raise NotImplementedError()
-
-
-class StatsUpdateBase(object):
-    @abc.abstractmethod
-    def update_stats(self, health_message, srcaddr):
-        raise NotImplementedError()
